% Generated by roxygen2 (4.0.2): do not edit by hand
\name{acor_table}
\alias{acor_table}
\title{Calculates correlogram values on a raster using fft.}
\usage{
acor_table(x, maxdist = 1500, verbose = F)
}
\arguments{
\item{x}{A raster* object. Missing values are indicated by NA.}

\item{maxdist}{Maximum distance (in km) to include in correlogram table.  All possible distances are calculated using FFT, then trimmed to this value.}
}
\value{
The table of distances
}
\description{
Applies the Wiener-Khinchin theorem to extract spatial autocorrelation using Fast Fourier Transform techniques.  This results in an extremely fast way to calculate a complete correlogram (correlation as a function of distance) for a raster image.
}
\references{
\url{en.wikipedia.org/wiki/WienerKhinchin_theorem}

Xianlin Ma, Tingting Yao, A program for 2D modeling (cross) correlogram tables using fast Fourier transform, Computers & Geosciences, Volume 27, Issue 7, August 2001, Pages 763-774, ISSN 0098-3004, \url{http://dx.doi.org/10.1016/S0098-3004(01)00007-3}.

\url{http://www.johnloomis.org/ece563/notes/freq/autoself/autoself.htm}

\url{http://www.seas.upenn.edu/~ese502/NOTEBOOK/Part_II/3_Spatially_Dependent_Random_Effects.pdf}
}

