% Generated by roxygen2 (4.0.1): do not edit by hand
\name{acorr_center}
\alias{acorr_center}
\title{Finds the center pixel to faciltate comparison of results from \code{acorr}.}
\usage{
acorr_center(x)
}
\arguments{
\item{x}{A raster* object}
}
\value{
A raster object with all cells NA except the center pixel.
}
\description{
Finds the center pixel to faciltate comparison of results from \code{acorr}.
}

