% Generated by roxygen2 (4.0.2): do not edit by hand
\name{acorr_dist}
\alias{acorr_dist}
\title{Calculates the distance of each cell in a raster to the center pixel to faciltate comparison of results from \code{acorr} as a function of distance.}
\usage{
acorr_dist(x)
}
\arguments{
\item{x}{A raster* object}
}
\value{
A raster object showing the distances from the center pixel
}
\description{
Calculates the distance of each cell in a raster to the center pixel to faciltate comparison of results from \code{acorr} as a function of distance.
}

