% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcorr2.R
\name{acorr_xcorr2}
\alias{acorr_xcorr2}
\title{Use Octave's \code{xcorr2} to get spatial autocorrelation of complete image}
\usage{
acorr_xcorr2(x)
}
\arguments{
\item{x}{A raster* object}
}
\value{
A raster object of the autocorrelation of the image
}
\description{
Use Octave's \code{xcorr2} to get spatial autocorrelation of complete image
}
\references{
\url{http://octave.sourceforge.net/signal/function/xcorr2.html}
}

