% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fftshift2}
\alias{fftshift2}
\title{Rearranges outputs from fft to move the zero-frequency component to the center of the matrix.  This is useful to visualize a Fourier transform with the zero-frequency component in the center of the image.}
\usage{
fftshift2(x)
}
\arguments{
\item{x}{A matrix returned from \code{fft()}}
}
\value{
The transformed matrix
}
\description{
Rearranges outputs from fft to move the zero-frequency component to the center of the matrix.  This is useful to visualize a Fourier transform with the zero-frequency component in the center of the image.
}
\references{
\url{http://www.mathworks.com/help/matlab/ref/fftshift.html}

\url{http://stackoverflow.com/questions/5735720/effcient-way-to-do-fft-shift-in-matlab-without-using-fftshift-function}
}

