% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmRaster.R
\name{rmr}
\alias{rmr}
\title{Really Remove Raster Temp files.}
\usage{
rmr(x)
}
\arguments{
\item{x}{A raster* object.}
}
\description{
Really Remove Raster Temp files. This function checks the path of a raster object and removes it from disk as well as from the R workspace.
}

