% Generated by roxygen2 (4.0.1): do not edit by hand
\name{acorr_dir}
\alias{acorr_dir}
\title{Calculates the direction of each cell in a raster to the center pixel to faciltate comparison of results from \code{acorr} as a function of distance.}
\usage{
acorr_dir(x)
}
\arguments{
\item{x}{A raster* object}
}
\value{
A raster object showing the direction (in degrees) from the center pixel
}
\description{
Calculates the direction of each cell in a raster to the center pixel to faciltate comparison of results from \code{acorr} as a function of distance.
}

