% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifft.R
\name{ifft}
\alias{ifft}
\title{Calculates the normalized inverse fft of an array}
\usage{
ifft(x)
}
\arguments{
\item{x}{An array as used by fft()}
}
\value{
The normalized inverse fft of x
}
\description{
Convenience function to calculates the normalized inverse fft of an array (like in MatLab)
}
\references{
\url{http://stackoverflow.com/questions/8162562/is-there-a-package-in-r-that-gives-normalized-inverse-fft}
}

