% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rmr}
\alias{rmr}
\title{Calculates the full spatial autocorrelation on a raster using fft.}
\usage{
rmr(x)
}
\arguments{
\item{x}{A raster* object.}
}
\description{
Really Remove Raster Temp files. This function checks the path of a raster object and removes it from disk as well as from the R workspace.
}

